refactor: use .exo.ts convention for EMH
