+++ packages/portfolio-contract/test/snapshots/evm-wallet-handler.exo.test.ts.md
@@ -1,6 +1,6 @@
-# Snapshot report for `test/evm-wallet-handler.test.ts`
+# Snapshot report for `test/evm-wallet-handler.exo.test.ts`
 
-The actual snapshot is saved in `evm-wallet-handler.test.ts.snap`.
+The actual snapshot is saved in `evm-wallet-handler.exo.test.ts.snap`.
 
 Generated by [AVA](https://avajs.dev).
